pyproject.toml
[tool.poetry]
name = "claim-assistant"
version = "0.1.0"
description = "Construction Claim Assistant CLI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "claimctl"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
rich = "^13.7.0"
openai = "^1.12.0"
faiss-cpu = "^1.7.4"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
pymupdf = "^1.23.0"
pytesseract = "^0.3.10"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"
ruff = "^0.1.3"

[tool.poetry.scripts]
claimctl = "claimctl.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = []
